+++ src/app/(dashboard)/settings/layout.js
@@ -1,19 +1,32 @@
 "use client";
 
+import { useState } from "react";
 import Link from "next/link";
-import { usePathname } from "next/navigation";
+import { usePathname, useRouter } from "next/navigation";
 import { useUser } from "@/context/UserContext";
 import useAuth from "@/hooks/useAuth";
+import API from "@/api/api";
 
 import Container from "@/components/Container";
 import Subheading from "@/components/Subheading";
+import Modal from "@/components/Modal";
+import Button from "@/components/Button";
+import Form from "@/components/Form";
 
 const Layout = ({ children }) => {
+  const [error, setError] = useState({});
+  const [success, setSuccess] = useState({});
+  const [isOpen, setIsOpen] = useState(false);
+  const [loading, setLoading] = useState(false);
+  const [email, setEmail] = useState("");
+
   const pathname = usePathname();
   const { user } = useUser();
+  const { token } = useAuth();
+  const router = useRouter();
   const { subscribed } = useAuth();
 
-  const DEFAULT_LINKS = [
+  const LINKS = [
     {
       href: "/settings",
       label: "General",
@@ -26,39 +39,62 @@ const Layout = ({ children }) => {
       href: "/settings/password",
       label: "Password",
     },
+    ...(user && subscribed
+      ? [
+          {
+            href: "/settings/subscription",
+            label: "Subscription",
+          },
+        ]
+      : []),
+    ...(user && user.role === "super_admin"
+      ? [
+          {
+            href: "/settings/company",
+            label: "Edit Company Info",
+          },
+        ]
+      : []),
   ];
 
-  const LINKS = {
-    creator: [
-      ...DEFAULT_LINKS,
-      ...(user && subscribed
-        ? [
-            {
-              href: "/settings/subscription",
-              label: "Subscription",
-            },
-          ]
-        : []),
-    ],
-    brand: [
-      ...DEFAULT_LINKS,
-      ...(user && subscribed
-        ? [
-            {
-              href: "/settings/subscription",
-              label: "Subscription",
-            },
-          ]
-        : []),
-    ],
-    admin: [...DEFAULT_LINKS],
-    super_admin: [
-      ...DEFAULT_LINKS,
-      {
-        href: "/settings/company",
-        label: "Edit Company Info",
-      },
-    ],
+  const handleDeleteAccount = async () => {
+    setLoading(true);
+    setError({});
+    setSuccess({});
+    try {
+      const response = await API.delete("/auth/delete-account", {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+        data: {
+          email,
+        },
+      });
+
+      if (response.status === 200) {
+        setSuccess({
+          message: "Account deleted successfully",
+        });
+        router.push("/login");
+      } else {
+        setError({
+          message: response.data.message || "Failed to delete account",
+        });
+      }
+    } catch (error) {
+      setError({
+        message:
+          error.response?.data.message ||
+          error?.message ||
+          "Failed to delete account. Please try again later.",
+      });
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const handleemailChange = (event) => {
+    setEmail(event.target.value);
   };
 
   return (
@@ -71,7 +107,7 @@ const Layout = ({ children }) => {
                 Settings
               </Subheading>
               <ul>
-                {LINKS[user?.role]?.map(({ href, label }) => (
+                {LINKS.map(({ href, label }) => (
                   <li key={href} className="py-1">
                     <Link
                       href={href}
@@ -83,13 +119,51 @@ const Layout = ({ children }) => {
                 ))}
               </ul>
             </div>
-            <button type="button" className="text-red-600">
+            <button
+              type="button"
+              className="text-red-600"
+              onClick={() => setIsOpen(true)}
+            >
               Delete account
             </button>
           </div>
           <div className="w-full">{children}</div>
         </div>
       </div>
+
+      <Modal
+        open={isOpen}
+        onClose={() => setIsOpen(false)}
+        className="min-h-64 max-w-2xl"
+      >
+        <Subheading size="lg">Confirm account deletion</Subheading>
+
+        <Form handleSubmit={handleDeleteAccount} setError={setError}>
+          <div className="space-y-6">
+            <p className="mb-4">
+              Please type your email to confirm account deletion. This action
+              cannot be undone.
+            </p>
+            <input
+              type="text"
+              value={email}
+              onChange={handleemailChange}
+              className="border border-gray-300 rounded px-3 py-1 w-full mt-1"
+              required
+            />
+            <Button
+              as="button"
+              type="submit"
+              className="mt-8"
+              disabled={!email.trim()}
+            >
+              {loading && <Button.Spinner />} Confirm
+            </Button>
+          </div>
+          {error?.message && <Form.Error>{error.message}</Form.Error>}
+          {success?.message && <Form.Success>{success.message}</Form.Success>}
+        </Form>
+      </Modal>
     </Container>
   );
 };
